step 1 of partial merge from mdr2 branch - don't render natural=coastline as lines, this is rarely useful - don't fill string table and don't create tempFile for strings when calculating global index for gnapsupp. This reduces I/O and peak memory usage.
git-svn-id: https://svn.mkgmap.org.uk/mkgmap/mkgmap@4896 25d90789-57f7-4ee0-8453-03a3dfeeeb22

+++ trunk/src/uk/me/parabola/imgfmt/app/mdr/Mdr15.java
@@ -49,7 +49,11 @@ public class Mdr15 extends MdrSection {
 		setConfig(config);
 
 		charset = config.getSort().getCharset();
-
+		if (config.isForDevice()) {
+			tempFile = null;
+			stringFile = null;
+			return;
+		}
 		try {
 			tempFile = File.createTempFile("strings", null, config.getOutputDir());
 			tempFile.deleteOnExit();
@@ -98,7 +102,9 @@ public class Mdr15 extends MdrSection {
 	public void releaseMemory() {
 		strings = null;
 		try {
-			stringFile.close();
+			if (stringFile != null) {
+				stringFile.close();
+			}
 		} catch (IOException e) {
 			throw new MapFailedException("Could not close string temporary file");
 		}
